requirements.txt
asgiref==3.5.2
astroid==2.12.10
automium==0.2.6
automium-web==0.1.1
certifi==2022.9.24
chardet==5.0.0
click==8.1.3
Django==4.1.2
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
# idna==2.8
isort==5.10.1
itsdangerous==2.1.2
lazy-object-proxy==1.7.1
legacy==0.1.6
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
parts==0.2.1
Pillow==9.2.0
pytz==2020.4
requests==2.22.0
six==1.13.0
sqlparse==0.4.1
# sslcommerz-python==0.0.7
# typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.11.2
